fix(daemon): reconcile structural split contracts
