+++ packages/daemon/src/agent-runtime-launch-config.ts
@@ -1,93 +1,24 @@
-import { createHash } from "node:crypto";
-import {
-  accessSync,
-  chmodSync,
-  constants,
-  copyFileSync,
-  existsSync,
-  lstatSync,
-  mkdirSync,
-  readFileSync,
-  readlinkSync,
-  realpathSync,
-  renameSync,
-  rmSync,
-  symlinkSync,
-  writeFileSync,
-} from "node:fs";
-import { userInfo } from "node:os";
 import path from "node:path";
 import { consumeKnownError } from "../../kernel/src/index.ts";
-import type { AgentDefinitionSnapshot } from "../../kernel/src/index.ts";
-import { credentialPort, isCredentialReferenceText } from "./agent-runtime-credential-port.ts";
-import type { AgentRuntimeInstanceDto } from "./agent-runtime-contract.ts";
 import {
   permissionLaunchArgs,
-  runtimeIsolationState,
-  runtimePermissionMode,
   type RuntimeIsolationState,
   type RuntimePermissionMode,
 } from "./runtime-permissions.ts";
 import type {
   RuntimeInstanceKind,
-  RuntimeInstanceAuth,
-  RuntimeInstanceCommon,
-  ClaudeRuntimeInstanceConfig,
-  CodexRuntimeInstanceConfig,
-  AgyRuntimeInstanceConfig,
   RuntimeInstanceConfig,
   RuntimeInstallationWitness,
   RuntimeAuthReadiness,
-  RuntimeInstanceSummary,
-  PreparedRuntimeLaunch,
-  PreparedRuntimeAuthCommand,
 } from "./agent-runtime-instance-types.ts";
+import { runExecutableSync } from "./agent-runtime-installation-discovery.ts";
+import { providerConfigDirectory, tomlString } from "./agent-runtime-instance-storage.ts";
 import {
-  runtimeModelCatalogCache,
-  discoverRuntimeInstallations,
-  discoverRuntimeModelCatalog,
-  versionProbeEnvironment,
-  runExecutableSync,
-} from "./agent-runtime-installation-discovery.ts";
-import { openRuntimeInstanceStore } from "./agent-runtime-instance-store.ts";
-import {
-  migrateLegacyInstallationIdentities,
-  requireWitnessedInstallation,
-  missingInstallationHint,
-  installationLabel,
-  publicConfig,
-  definitionSnapshot,
-  providerConfigDirectory,
-  providerAuthFile,
-  sharedProviderDirectory,
-  ensureSharedAuthFile,
-  runtimeBaseUrl,
-  writeCodexConfig,
-  tomlString,
-} from "./agent-runtime-instance-storage.ts";
-import {
-  runtimeInstanceConfig,
-  claudeRuntimeConfig,
-  codexRuntimeConfig,
-  agyRuntimeConfig,
-  runtimeHttpHeaders,
-  needsRuntimeInstanceNormalization,
-  normalizeModels,
-  requireBoolean,
-  selectRuntimeModel,
-  selectRuntimeEffort,
-  runtimeEffort,
-  agyEffort,
-  runtimeInstanceId,
-  identifier,
-  credentialReference,
   requiredRuntimeInstanceText,
-  isRuntimeInstanceRecord,
   runtimeInstanceError,
   available,
   unavailable,
 } from "./agent-runtime-instance-config.ts";
-import type { LegacyRuntimeInstanceConfig, FlatRuntimeInstanceConfig } from "./agent-runtime-instance-config.ts";
 
 // Platform-derived isolation environment. POSIX keeps HOME/TMPDIR/XDG_RUNTIME_DIR
 // semantics; Windows derives the same guarantees from USERPROFILE/TEMP/APPDATA and
